+++ packages/wordpress/CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Bug Fixes
 - Fixed a bug in the front end where the style engine was not being flushed and rewritten when the block content was created by blockera site builder.
 
+### Improvements
+- Improve admin page design.
+
 ## 1.1.1 (2025-04-12)
 
 ### Improvements